feat(windstyle): Add material color plugin
This commit adds a new file `material-color.ts` to the `windstyle` package. The file contains a plugin for Tailwind CSS that adds utility classes for material color scheme. It includes utility classes for light, dark, and auto color schemes, as well as media rule colors for dark mode. The `convertHexColorToRGBA` function is also included to convert hex color codes to RGBA format.